lotsa refactors
