+++ src/client/src/assets/CropIcons/bamboo.png
