+++ src/client/src/assets/CropIcons/potato.png
