+++ src/client/src/components/Shop/constants.js
@@ -1,0 +1,44 @@
+export const shopItemsList = [
+    {
+      id: 1,
+      name: "Corn",
+      price: 100,
+      // from https://www.flaticon.com/free-icons/corn
+      image: require("../../assets/CropIcons/corn.png"),
+    },
+    {
+      id: 2,
+      name: "Mushroom",
+      price: 200,
+      // from https://www.flaticon.com/free-icons/mushroom
+      image: require("../../assets/CropIcons/mushroom.png"),
+    },
+    {
+      id: 3,
+      name: "Lavender",
+      price: 150,
+      // from https://www.flaticon.com/free-icons/flower
+      image: require("../../assets/CropIcons/lavender.png"),
+    },
+    {
+      id: 4,
+      name: "Bamboo",
+      price: 150,
+      // from https://www.flaticon.com/free-icons/bamboo
+      image: require("../../assets/CropIcons/corn.png"),
+    },
+    {
+      id: 5,
+      name: "Wheat",
+      price: 125,
+      // from https://www.flaticon.com/free-icons/rice
+      image: require("../../assets/CropIcons/wheat.png"),
+    },
+    {
+      id: 6,
+      name: "Potato",
+      price: 175,
+      // from https://www.flaticon.com/free-icons/potato
+      image: require("../../assets/CropIcons/potato.png"),
+    },
+  ];
