+++ src/client/src/components/models/TestModel.js
@@ -1,4 +1,4 @@
-import React, { Suspense } from "react";
+import React from "react";
 import { GLTFLoader } from 'three/examples/jsm/loaders/GLTFLoader';
 import { useLoader } from "@react-three/fiber";
